Introduce interfaces and classes for built-in thumbnailers.
This commit introduces the TumblerThumbnailer interface which is
supposed to be implemented twice, once for built-in thumbnailers that
run in the same process as tumbler itself, and once for specialized
thumbnailers which talk to tumbler over D-Bus. TumblerThumbnailer
provides a function for creating a thumbnail plus "ready" and "error"
signals. Their signatures may still change though.

TumblerBuiltinThumbnailer is a new class for built-in thumbnailers. It
is a little similar to ExoSimpleJob in that it allows built-in
thumbnailers to be created by defining a create function and a list of
mime types.

Also new is tumbler_pixbuf_thumbnailer_new() which creates a pixbuf
thumbnailer. It's just a skeleton for now.

TumblerRegistry now has a method tumbler_registry_add() for adding a
new TumblerThumbnailer to the registry.

Last but not least, there's tumbler-marshal.list for custom marshallers
now.