Improve TumblerThumbnailer and add TumblerSpecializedThumbnailer.
Add new properties "uri-schemes" and "hash-keys" to TumblerThumbnailer
and add getter functions for all TumblerThumbnailer properties.

Add URI schemes parameter to the TumblerBuiltinThumbnailer constructor
and compute hash keys for built-in thumbnailers based on the provided
URI schemes and mime types.

Adds TumblerSpecializedThumbnailer which is created for each thumbnailer
either found in $XDG_DATA_DIRS/thumbnailers/ or registered dynamically
via org.freedesktop.thumbnailer.Manager.Register(). Internally it uses a
DBusGProxy to talk to the specialized thumbnailer service.

Add a GMutex to TumblerManager and lock all functions for now. Implement
the Register() method which creates a new TumblerSpecializedThumbnailer
and adds it to the registry.