Implement thumbnailer management in TumblerRegistry.
Use a (gchar *) to (GList **) hash table for managing thumbnailers in
TumblerRegistry. For each so-called hash key (a combination of an URI
scheme and a MIME type, e.g. "file-image/png") we store a list of
registered thumbnailers. Only the first thumbnailer in this list will
ever be used. The others can move forward in the list when a thumbnailer
is unregistered.

The order in which thumbnailers are inserted into a list is defined as
follows:

  - thumbnailers registered dynamically over D-Bus go before anything
    else. The thumbnailer registered latest goes first.
  - specialized thumbnailers with most recent service files go first.
  - built-in thumbnailers are always at the end of the list in the
    order they were added to the registry.

Connect to the new "unregister" signal of all TumblerThumbnailers to
remove them when required. The signal is only emitted by
TumblerSpecializedThumbnailers when their proxy is destroyed.

TumblerSpecializedThumbnailer has two new properties: "foreign" and
"modified". Foreign thumbnailers are those registered dynamically over
D-Bus (using org.freedesktop.thumbnailer.Manager.Register()). The
"modified" property represents the timestamp of when the thumbnailer's
.service file was last modified. It is 0 for foreign thumbnailers.