Add TumblerScheduler and TumblerThresholdScheduler skeletons.
Thumbnail requests need to be scheduled in some way. Let's say the
service receives a request for 1000 thumbnails and one for 5 after the
first one. How should these be scheduled?

TumblerScheduler provides an interface for different schedulers, one of
which is TumblerThresholdScheduler. TumblerThresholdScheduler is
inspired by the thumbnailer in hildon-thumbnail. It uses two thread
pools, one for large and one for small requests. Requests are managed
using a LIFO and inserted into one of the pools depending on the number
of requested thumbnails (the threshold).

However, for the moment these are just skeletons.

This commit also makes sure to copy hash keys before inserting
thumbnailers into the registry hash table.