Implement TumblerNaiveScheduler and add signals to TumblerService.
TumblerNaiveScheduler is a very simple scheduler running in the main
loop. It processes requests as they come and emits signals per-URI. It
is mainly useful for testing purposes.

With the first scheduler being available, implement "started",
"finished", "error" and "ready" signals in TumblerThumbnailer,
TumblerBuiltinThumbnailer, TumblerScheduler, TumblerNaiveScheduler and
TumblerService.

Also implement tumbler_builtin_scheduler_create() based on the
TumblerBuiltinThumbnailerFunc specified in the constructor. We now have
an environment to implement and test the built-in pixbuf thumbnailer.