No thumbnails for thumbnails. Allow Unqueue in the middle of requests.
There's a new method tumbler_cache_is_thumbnail() now which has to
be implemented by plugins. It is used by TumblerFileInfo in
tumbler_file_info_load() to check whether the file is a thumbnail
itself. In that case, tumbler_file_info_load() will fail with
a TUMBLER_ERROR_IS_THUMBNAIL error.

Add another check in TumblerThresholdScheduler to make unqueuing
requests in the middle of the process work.