Add tumbler_thumbnail_save_file() to save thumbnails from source files.
This is especially useful for plugins that execute external commands
such as "convert". They can just let these commands write to a certain
file which they then pass to tumbler_thumbnail_save_file() to make sure
it's stored in a way the corresponding cache plugin supports.

It is an alternative to tumbler_thumbnail_save_pixbuf().