Back to the org.freedesktop.thumbnails namespace. Use API versioning.
This is another in a series of commits aiming at synchronizing the
implemented D-Bus API with the changes recently made to the spec on
http://live.gnome.org/ThumbnailerSpec.

We now append an API version to the D-Bus interface and object path
names. We also use org.freedesktop.thumbnails.SpecializedThumbnailer for
specialized thumbnailers, as we do not need to use another namespace
(previously org.freedesktop.thumbnailer.Thumbnailer was used) for it.