Optimize tumbler_service_get_supported() by not duplicating strings.
Instead, we remember all strings used in the process of generating URI
scheme / MIME type pairs. We don't free them immediately but leave them
in memory for reuse in the hash table used to collect unique pairs and
in the pair arrays returned by the GetSupported() D-Bus method. We free
them at the very end of the function.