Switch to a plugin API that only allows one cache backend at a time.
Tumbler will now only check for the file

  $(libdir)/tumbler-1/plugins/cache/tumbler-cache-plugin.so

which is supposed to link to the cache backend that is being used. If
the XDG cache is built, tumbler-cache-plugin.so is set up to link to
this plugin.

The fact that we only have a singleton cache makes things much less
ambiguous.