Allow TumblerFileInfos to be created with empty URIs/MIME types.
This is done to avoid warnings when someone forgets to send a MIME type
along with a thumbnail request etc. For empty URIs/MIME types, tumbler
will not find a thumbnailer and the broken elements will be reported
back to the client via the "Error" signal.

The original idea came from Philip Van Hoof <philip@codeminded.be>.