Make Cleanup() take an array of base URIs instead of just one.
The thumbnail management D-Bus specification was updated accordingly.

This change makes sense due to the following reason: when trashing a
directory we cannot simply call org.freedesktop.thumbnails.Cache1.Delete
as this will only remove the thumbnail of the directory itself, not of
its descendants. But we can use the Cleanup method for this. With
Cleanup taking an array of base URIs instead of just one we can avoid
hundreds of Cleanup calls when trashing a large number of files.