First try at a webkit thumbnailer.
It does not work here because GtkOffscreenWindow crashes when shown.
Seems to be a Gtk bug which I'm investigating. I don't guarantee that it
will work for anyone as I obviously could not test it. For example, I
don't know what will happen when called on a huge HTML page.