Make the thumbnailer actually work.
Set the load-finished callback before starting to load an uri and
actually show the web view so that we don't get white thumbnails...

Remove the damage callback as it makes everything lag. Instead force a
window update before retrieving the pixbuf.

Also set a size request for the WebKitView so that it's not 1x1. For
now default to a classic netbook size, needs further thinking.