Handle 0 mtime threshold on Cleanup DBus method.
According to the Thumbnail management DBus specification,
original_mtime_threshold can be 0 to ignore the threshold and only
cleanup based on the URI prefix array. Previously this was ignored,
on 'Cleanup' method call, tumbler was running across all thumbnails
in the thumbnail directory, using a lot of I/O bandwidth on any file
deleted by Thunar that has a thumbnail, bug #9231.