Accept network paths in ffmpeg thumbnailer plugin
g_file_get_path() appears to be able to convert smb://, sftp://, and
other GVFS schemes to local absolute paths which can be passed as is
to ffmpegthumbnailer. In order to handle these URIs, remove the call
to g_file_is_native() and rely exclussively on g_file_get_path().