pixbuf-thumbnailer: Use a mime-type-specific loader when possible
This allows to fully use the capabilities of the specific pixbuf loader
for the considered mime type according to the installed optional
libraries. This is especially necessary to load some SVG or HEIF
thumbnails.

On the other hand, if this specific loader doesn't work properly, it may
be even more noticeable. This is the case for libopenraw, for which a
dedicated plugin using another API has fortunately been implemented. So
we just have to prioritize it in `tumbler.rc`.

Fixes #20, #32, #36.

Co-authored-by: @Krifa75 (Gitlab ID)