Fix `-Wclobbered` compilation warnings
These warnings appear only under certain conditions and/or for certain
versions of the compilers. For example with GCC 11.1 when building with
`--disable-debug` and with Clang 13.0.

They had been left out in !16 with doubts about their validity, but they
are actually understandable in view of `setjmp()` calls, and can be
fixed easily.