Revert "Set Max file sizes."
This reverts commit 8f0234a4a14bfec9f0a1ef5c1c8b94fb523f5d41.

There was no particular justification for this change, and since then
there have been several interventions in bug reports or on the forum
from users wondering why Tumbler does not generate thumbnails for
certain files, especially video.

There should not be a default size limitation, or only for a particular
plugin, with a justification in the comments, and only if the problem
cannot be fixed in the code.