Try all available thumbnailers before declaring failure
The request thumbnailer array no longer contains one thumbnailer per
URI, but the list of available thumbnailers for this URI (after applying
the filters of the configuration file), sorted by priority.

For a given URI, we browse the thumbnailer list until we succeed in
generating a thumbnail. An error signal is only issued if the last
thumbnailer has failed. Intermediate errors are only displayed if debug
logging is enabled.

Closes #50.