Store modification time in locale independent format
A mistake in a242e1324c7a3d1428e2a8ce58ee77c71967d8b7: the `printf()`
function family is obviously locale dependent. So either avoid it or use
`strtod()` for reverse conversion, which is also locale dependent.
Clearly it is best to remain locale independent for this kind of
information.

Related: #15, !20