Yet another fix to the modification time storage format
Another option would have been to keep separate guint64 and guint32
storage internally for the integer and fractional part of the
modification time, shifting the difficulty to string manipulations when
restoring from disk. Ultimately, this may be safer, but it is certainly
more cumbersome and should work that way.

Related: #15, !20