desktop-thumbnailer: Guard against null path
It can for example be null for a file in the trash can. The working
directory that follows is optional, and the path is only required
afterwards if the thumbnailer command uses the `%i` field code, instead
of `%u`. In this case, the command will fail in its own way, most likely
gracefully.