desktop-thumbnailer: Use empty path to build cmdline if it is null
This is probably best in general, as otherwise adding the path is simply
ignored in desktop_thumbnailer_exec_parse(), which has the effect of
shifting the command line parameters. It's then up to the desktop
thumbnailer executable to correctly handle the case of an empty path,
but at least it will have the right number of parameters.

Completes: 1719ce1a47003bb755e63da6f6f3a0dee12855b7