scheduler: Fix ready/error signal forwarding
Since 9980fc3b, multiple requests can take place in parallel, although
URIs within a single request are still processed sequentially. As
thumbnailers are common to all requests, the ready and error signal
for thumbnailers need to be filtered to ensure that a signal is only
forwarded if it relates to the right request.

On the one hand, this fixes duplicate signals, but it may also fix
partial request cancellations, as in some cases g_cancellable_cancel()
could be called on the ready signal for the wrong request (if a
thumbnailer succeeds for a URI, the invocation of lower-priority
thumbnailers for this URI is cancelled).

Fixes: 9980fc3b8d5043f2079b4a631caf7ca23a43d4fe
(cherry picked from commit 6e45f64e9a2c04301074b6a99bbb0e970562a2e8)