Disambiguate thumbnailer signals
Thumbnailer signals contain only the URI as original information,
whereas the same URI can be processed for different flavors or even
different mime types. This isn't a problem as long as requests are
processed sequentially, but it's not necessarily the case since
9980fc3b. More generally, these signals should return the
TumblerFileInfo that the thumbnailers receive when a request is made, to
remove any ambiguity when the scheduler filters the signals.

Completes: 6e45f64e9a2c04301074b6a99bbb0e970562a2e8